fix(db): connect standalone scripts + backup over ProxySQL-enforced TLS
ProxySQL now enforces SSL/TLS, so the nightly snapshot job failed with
"Access denied ... SSL is required" / pool timeout. The app runtime already
connected with SSL via buildDatabaseConfig() in server/utils/prisma.ts, but
the standalone maintenance scripts instantiated `new PrismaMariaDb(url)` with
no ssl config, and the fallback-snapshot workflow never passed the CA secret.

- Extract the SSL-aware adapter config builder out of server/utils/prisma.ts
  into a side-effect-free server/utils/databaseAdapterConfig.ts and reuse it
  (no duplication). prisma.ts imports buildDatabaseConfig from it unchanged.
- Route snapshotFallback.ts, exportRewards.ts, backfillSlugs.ts and
  repairCharacterSlugs.mjs through createDatabaseAdapter() so they complete
  the same CA-verified TLS handshake the app uses. The snapshot script keeps
  its 30s connectTimeout (preserved through the shared builder) and logs
  whether TLS is engaged (boolean only, no host/creds).
- fallback-snapshot.yml: pass DATABASE_SSL_CA_BASE64 into the snapshot and
  dump job envs; make the mysqldump backup and the connectivity-probe mysql
  CLI present the CA (--ssl-mode=VERIFY_CA); warn if the secret is absent.

TLS verification is NOT disabled (rejectUnauthorized stays true / VERIFY_CA).
davinci-seed-verify.yml is unaffected — it targets a local mariadb service
container, not the home DB.
